create types.ts, write interface for db schemas
